+++ packages/plugin-divider/web/src/getToolbarButtons.tsx
@@ -1,11 +1,23 @@
 import React from 'react';
 import type { ToolbarButton } from 'ricos-types';
 import { PLUGIN_TOOLBAR_BUTTON_ID } from 'wix-rich-content-editor-common';
-import { NodeAlignmentButton } from 'wix-rich-content-toolbars-ui';
 import type { PluginContainerData_Alignment } from 'ricos-schema';
+import { DividerAlignmentButton } from './toolbar/DividerAlignmentButton';
 import { DividerSizeButton } from './toolbar/DividerSizeButton';
 import { DividerStyleButton } from './toolbar/DividerStyleButton';
 import { TIPTAP_DIVIDER_TYPE } from 'ricos-content';
+import { DIVIDER_SIZES } from './toolbar/dividerButtonsData';
+
+const selectedNodeResolver = {
+  id: 'selectedNode',
+  resolve: content => {
+    if (Array.isArray(content) && content.length > 0) {
+      return content[0];
+    } else {
+      return undefined;
+    }
+  },
+};
 
 export const getToolbarButtons = (config, services): ToolbarButton[] => {
   return [
@@ -30,7 +42,16 @@ export const getToolbarButtons = (config, services): ToolbarButton[] => {
       id: 'dividerSize',
       dataHook: 'dividerSizeDropdownButton',
       command: ({ size, editorCommands }) => {
-        editorCommands.chain().focus().setNodeSize(size).setNodeAlignment('CENTER').run();
+        editorCommands
+          .chain()
+          .focus()
+          .updateAttributes(TIPTAP_DIVIDER_TYPE, {
+            width: DIVIDER_SIZES[size].toUpperCase(),
+          })
+          .run();
+      },
+      attributes: {
+        selectedNode: selectedNodeResolver,
       },
       renderer: toolbarItem => <DividerSizeButton toolbarItem={toolbarItem} />,
     },
@@ -38,9 +59,23 @@ export const getToolbarButtons = (config, services): ToolbarButton[] => {
       id: PLUGIN_TOOLBAR_BUTTON_ID.SEPARATOR,
     },
     {
-      id: PLUGIN_TOOLBAR_BUTTON_ID.ALIGNMENT,
+      id: 'dividerAlignment',
+      command: ({ alignment, editorCommands, node }) => {
+        const nodeContainerData = node.attrs?.containerData;
+        editorCommands
+          .chain()
+          .focus()
+          .updateAttributes(TIPTAP_DIVIDER_TYPE, {
+            alignment,
+            containerData: { ...nodeContainerData, alignment },
+          })
+          .run();
+      },
+      attributes: {
+        selectedNode: selectedNodeResolver,
+      },
       renderer: toolbarItem => (
-        <NodeAlignmentButton
+        <DividerAlignmentButton
           toolbarItem={toolbarItem}
           options={['LEFT', 'CENTER', 'RIGHT'] as PluginContainerData_Alignment[]}
         />